Added basic combat system

+++ Assets/Scripts/Game/Units/Spells/OnHits/BonusDamageOnHit.cs
@@ -5,9 +5,9 @@ namespace Game.Units.Spells.OnHits
 	[Serializable]
 	public class BonusDamageOnHit : OnHit
 	{
-		public int damage;
+		public float damage;
 
-		protected override int? ApplyEffect (int baseDamage, Game.Units.UnitBehaviour target)
+		protected override float? ApplyEffect (float baseDamage, Game.Units.UnitBehaviour target)
 		{
             return baseDamage + damage;
 		}
