+++ Assets/Scripts/Game/Units/Spells/OnHits/DotOnHit.cs
@@ -4,7 +4,7 @@ namespace Game.Units.Spells.OnHits
 {
 	public class DotOnHit : OnHit
 	{
-		protected override int? ApplyEffect (int baseDamage, UnitBehaviour target)
+		protected override float? ApplyEffect (float baseDamage, UnitBehaviour target)
 		{
             throw new NotImplementedException();
 		}
